small cleanup before Rc push
